+++ src/ConfigProvider.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Ikoss\Mezzio\DebugBar;
 
-use DebugBar\Bridge\DoctrineCollector;
+use DebugBar\Bridge\Doctrine\DebugBarSQLMiddleware;
+use DebugBar\Bridge\Doctrine\DoctrineCollector;
 use DebugBar\DataCollector\ConfigCollector;
 use DebugBar\DebugBar;
 use DebugBar\Storage\FileStorage;
@@ -61,15 +62,16 @@ final class ConfigProvider
     {
         return [
             'factories' => [
-                DebugBarMiddleware::class => DebugBarMiddlewareFactory::class,
-                ConfigCollector::class    => ConfigCollectorFactory::class,
-                DoctrineCollector::class  => DoctrineCollectorFactory::class,
-                DebugBar::class           => StandardDebugBarFactory::class,
-                FileStorage::class        => FileStorageFactory::class,
-                OpenHandler::class        => OpenHandlerFactory::class,
-                DoctrineStorage::class    => DoctrineStorageFactory::class,
-                PdoStorage::class         => PdoStorageFactory::class,
-                RouteCollector::class     => RouteCollectorFactory::class,
+                DebugBarMiddleware::class    => DebugBarMiddlewareFactory::class,
+                ConfigCollector::class       => ConfigCollectorFactory::class,
+                DoctrineCollector::class     => DoctrineCollectorFactory::class,
+                DebugBar::class              => StandardDebugBarFactory::class,
+                FileStorage::class           => FileStorageFactory::class,
+                OpenHandler::class           => OpenHandlerFactory::class,
+                DoctrineStorage::class       => DoctrineStorageFactory::class,
+                PdoStorage::class            => PdoStorageFactory::class,
+                RouteCollector::class        => RouteCollectorFactory::class,
+                DebugBarSQLMiddleware::class => DebugBarSQLMiddlewareFactory::class,
             ],
         ];
     }