+++ src/DataCollector/RouteCollector.php
@@ -4,20 +4,21 @@ declare(strict_types=1);
 
 namespace Ikoss\Mezzio\DebugBar\DataCollector;
 
-use DebugBar\DataCollector\AssetProvider;
 use DebugBar\DataCollector\DataCollector;
 use DebugBar\DataCollector\Renderable;
+use DebugBar\DataCollector\Resettable;
 use Laminas\Diactoros\ServerRequestFactory;
 use Mezzio\Router\RouterInterface;
 
 use function is_string;
 
-class RouteCollector extends DataCollector implements Renderable, AssetProvider
+class RouteCollector extends DataCollector implements Renderable, Resettable
 {
     protected string $name;
 
     protected array $config;
 
+    protected array $data;
     protected RouterInterface $router;
 
     public function __construct(RouterInterface $router, array $config)
@@ -27,18 +28,23 @@ class RouteCollector extends DataCollector implements Renderable, AssetProvider
         $this->name   = 'Route';
     }
 
+    public function reset(): void
+    {
+        $this->data = [];
+    }
+
     public function collect(): array
     {
-        $data = $this->getRouteInformation();
-        foreach ($data as $k => $v) {
+        $this->data = $this->getRouteInformation();
+        foreach ($this->data as $k => $v) {
             if ($this->isHtmlVarDumperUsed()) {
-                $v = $this->getVarDumper()->renderVar($v);
+                $v = $this->getDataFormatter()->formatVar($v);
             } elseif (! is_string($v)) {
                 $v = $this->getDataFormatter()->formatVar($v);
             }
-            $data[$k] = $v;
+            $this->data [ $k ] = $v;
         }
-        return $data;
+        return $this->data;
     }
 
     /**
@@ -59,34 +65,22 @@ class RouteCollector extends DataCollector implements Renderable, AssetProvider
         return $this->config[ 'routes' ][ $match->getMatchedRouteName() ] ?? ['no data'];
     }
 
-    /**
-     * @return string
-     */
-    public function getName()
+    public function getName(): string
     {
         return $this->name;
     }
 
-    /**
-     * @return array
-     */
-    public function getAssets()
-    {
-        return $this->isHtmlVarDumperUsed() ? $this->getVarDumper()->getAssets() : [];
-    }
-
-    /**
-     * @return array
-     */
-    public function getWidgets()
+    public function getWidgets(): array
     {
         $name   = $this->getName();
-        $widget = $this->isHtmlVarDumperUsed()
-            ? "PhpDebugBar.Widgets.HtmlVariableListWidget"
-            : "PhpDebugBar.Widgets.VariableListWidget";
+        $widget = match (true) {
+            $this->isJsonVarDumperUsed() => "PhpDebugBar.Widgets.JsonVariableListWidget",
+            $this->isHtmlVarDumperUsed() => "PhpDebugBar.Widgets.HtmlVariableListWidget",
+            default => "PhpDebugBar.Widgets.VariableListWidget",
+        };
         return [
             "$name" => [
-                "icon"    => "gear",
+                "icon"    => "adjustments",
                 "widget"  => $widget,
                 "map"     => "$name",
                 "default" => "{}",