+++ tests/Storage/MockStorage.php
@@ -7,14 +7,12 @@ namespace Ikoss\Mezzio\DebugBar\Tests\Storage;
 use DebugBar\Storage\StorageInterface;
 
 use function array_slice;
+use function time;
 
 class MockStorage implements StorageInterface
 {
     public array $data;
 
-    /**
-     * @inheritDoc
-     */
     public function __construct(array $data = [])
     {
         $this->data = $data;
@@ -23,7 +21,7 @@ class MockStorage implements StorageInterface
     /**
      * @inheritDoc
      */
-    public function save($id, $data)
+    public function save(string $id, array $data): void
     {
         $this->data[$id] = $data;
     }
@@ -31,7 +29,7 @@ class MockStorage implements StorageInterface
     /**
      * @inheritDoc
      */
-    public function get($id)
+    public function get(string $id): array
     {
         return $this->data[$id];
     }
@@ -39,13 +37,32 @@ class MockStorage implements StorageInterface
     /**
      * @inheritDoc
      */
-    public function find(array $filters = [], $max = 20, $offset = 0)
+    public function find(array $filters = [], int $max = 20, int $offset = 0): array
     {
         return array_slice($this->data, $offset, $max);
     }
 
-    public function clear()
+    /**
+     * @inheritDoc
+     */
+    public function clear(): void
     {
         $this->data = [];
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function prune(int $hours = 24): void
+    {
+        $threshold = time() - $hours * 3600;
+
+        foreach ($this->data as $id => $entry) {
+            if (($entry['time'] ?? 0) >= $threshold) {
+                continue;
+            }
+
+            unset($this->data[$id]);
+        }
     }
 }